统计使用量接口更新

+++ src/main/java/com/berry/oss/module/dto/BucketStatisticsInfoDto.java
@@ -21,7 +21,7 @@ public class BucketStatisticsInfoDto {
     /**
      * object 个数
      */
-    private Integer objectCount;
+    private Long objectCount;
 
     /**
      * 最大 object size
